refactor: extract Form component logic into custom hooks
- Extract filter state management into useFilterState hook
- Extract form state management into useFormState hook
- Extract UI state management into useUIState hook
- Reduce Form component complexity from 650+ to 400 lines
- Improve maintainability and code organization
- No UI/design changes, internal structure only
